Rename the ToolchainCluster controller to ToolchainCluster_cache controller
