+++ controllers/toolchaincluster_cache/healthchecker_test.go
@@ -1,4 +1,4 @@
-package toolchaincluster
+package toolchaincluster_cache
 
 import (
 	"context"
