+++ controllers/toolchaincluster_cache/toolchaincluster_cache_controller.go
@@ -1,4 +1,4 @@
-package toolchaincluster
+package toolchaincluster_cache
 
 import (
 	"context"
